+++ src/main/java/io/passport/server/controller/LearningProcessController.java
@@ -1,13 +1,17 @@
 package io.passport.server.controller;
 
 import io.passport.server.model.LearningProcess;
+import io.passport.server.model.Role;
 import io.passport.server.service.LearningProcessService;
+import io.passport.server.service.RoleCheckerService;
+import org.keycloak.KeycloakPrincipal;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -25,17 +29,32 @@ public class LearningProcessController {
      */
     private final LearningProcessService learningProcessService;
 
+    /**
+     * Role checker service for authorization
+     */
+    private final RoleCheckerService roleCheckerService;
+
     @Autowired
-    public LearningProcessController(LearningProcessService learningProcessService) {
+    public LearningProcessController(LearningProcessService learningProcessService, RoleCheckerService roleCheckerService) {
         this.learningProcessService = learningProcessService;
+        this.roleCheckerService = roleCheckerService;
     }
 
     /**
      * Read all learning processes
+     * @param principal KeycloakPrincipal object that holds access token
      * @return
      */
     @GetMapping()
-    public ResponseEntity<List<LearningProcess>> getAllLearningProcesses() {
+    public ResponseEntity<List<LearningProcess>> getAllLearningProcesses(@AuthenticationPrincipal KeycloakPrincipal<?> principal) {
+
+        // Allowed roles for this endpoint
+        List<Role> allowedRoles = List.of(Role.DATA_SCIENTIST);
+        // Check role of the user
+        if(!this.roleCheckerService.hasAnyRole(principal, allowedRoles)){
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+
         List<LearningProcess> learningProcesses = this.learningProcessService.getAllLearningProcesses();
 
         long totalCount = learningProcesses.size();
@@ -49,10 +68,20 @@ public class LearningProcessController {
     /**
      * Read a learning process by id
      * @param learningProcessId ID of the learning process
+     * @param principal KeycloakPrincipal object that holds access token
      * @return
      */
     @GetMapping("/{learningProcessId}")
-    public ResponseEntity<?> getLearningProcess(@PathVariable Long learningProcessId) {
+    public ResponseEntity<?> getLearningProcess(@PathVariable Long learningProcessId,
+                                                @AuthenticationPrincipal KeycloakPrincipal<?> principal) {
+
+        // Allowed roles for this endpoint
+        List<Role> allowedRoles = List.of(Role.DATA_SCIENTIST);
+        // Check role of the user
+        if(!this.roleCheckerService.hasAnyRole(principal, allowedRoles)){
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+
         Optional<LearningProcess> learningProcess = this.learningProcessService.findLearningProcessById(learningProcessId);
 
         if(learningProcess.isPresent()) {
@@ -65,11 +94,21 @@ public class LearningProcessController {
     /**
      * Create LearningProcess.
      * @param learningProcess LearningProcess model instance to be created.
+     * @param principal KeycloakPrincipal object that holds access token
      * @return
      */
     @PostMapping()
-    public ResponseEntity<?> createLearningProcess(@RequestBody LearningProcess learningProcess) {
+    public ResponseEntity<?> createLearningProcess(@RequestBody LearningProcess learningProcess,
+                                                   @AuthenticationPrincipal KeycloakPrincipal<?> principal) {
         try{
+
+            // Allowed roles for this endpoint
+            List<Role> allowedRoles = List.of(Role.DATA_SCIENTIST);
+            // Check role of the user
+            if(!this.roleCheckerService.hasAnyRole(principal, allowedRoles)){
+                return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+            }
+
             LearningProcess savedLearningProcess = this.learningProcessService.saveLearningProcess(learningProcess);
             return ResponseEntity.status(HttpStatus.CREATED).body(savedLearningProcess);
         }catch(Exception e){
@@ -82,11 +121,22 @@ public class LearningProcessController {
      * Update LearningProcess.
      * @param learningProcessId ID of the learning process that is to be updated.
      * @param updatedLearningProcess LearningProcess model instance with updated details.
+     * @param principal KeycloakPrincipal object that holds access token
      * @return
      */
     @PutMapping("/{learningProcessId}")
-    public ResponseEntity<?> updateLearningProcess(@PathVariable Long learningProcessId, @RequestBody LearningProcess updatedLearningProcess) {
+    public ResponseEntity<?> updateLearningProcess(@PathVariable Long learningProcessId,
+                                                   @RequestBody LearningProcess updatedLearningProcess,
+                                                   @AuthenticationPrincipal KeycloakPrincipal<?> principal) {
         try{
+
+            // Allowed roles for this endpoint
+            List<Role> allowedRoles = List.of(Role.DATA_SCIENTIST);
+            // Check role of the user
+            if(!this.roleCheckerService.hasAnyRole(principal, allowedRoles)){
+                return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+            }
+
             Optional<LearningProcess> savedLearningProcess = this.learningProcessService.updateLearningProcess(learningProcessId, updatedLearningProcess);
             if(savedLearningProcess.isPresent()) {
                 return ResponseEntity.ok().body(savedLearningProcess);
@@ -102,11 +152,21 @@ public class LearningProcessController {
     /**
      * Delete by LearningProcess ID.
      * @param learningProcessId ID of the learning process that is to be deleted.
+     * @param principal KeycloakPrincipal object that holds access token
      * @return
      */
     @DeleteMapping("/{learningProcessId}")
-    public ResponseEntity<?> deleteLearningProcess(@PathVariable Long learningProcessId) {
+    public ResponseEntity<?> deleteLearningProcess(@PathVariable Long learningProcessId,
+                                                   @AuthenticationPrincipal KeycloakPrincipal<?> principal) {
         try{
+
+            // Allowed roles for this endpoint
+            List<Role> allowedRoles = List.of(Role.DATA_SCIENTIST);
+            // Check role of the user
+            if(!this.roleCheckerService.hasAnyRole(principal, allowedRoles)){
+                return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+            }
+
             boolean isDeleted = this.learningProcessService.deleteLearningProcess(learningProcessId);
             if(isDeleted) {
                 return ResponseEntity.noContent().build();
